fix: remove size_t for int comparison

+++ pointers_arrays_strings/8-print_array.c
@@ -1,5 +1,4 @@
 #include "main.h"
-#include <stddef.h>
 #include <stdio.h>
 /**
  * print_array - prints n elements
@@ -9,7 +8,7 @@
 void print_array(int *a, int n)
 {
 
-	size_t index = 0;
+	int index = 0;
 
 	while (index < n - 1)
 	{
